@@ -10,8 +10,8 @@ import java.math.BigDecimal;
 /**
  * Created by jt on 6/13/17.
  */
-@Getter
 @Setter
+@Getter
 @EqualsAndHashCode(exclude = {"recipe"})
 @Entity
 public class Ingredient {
